requirements.txt
spotipy==2.18.0
requests==2.25.1
requests_oauthlib==1.1.0
python3-discogs-client==2.3.12
